refactor(bin): rename file `event_queue` to `event_loop`
